% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_stan.R
\name{measles_stan}
\alias{measles_stan}
\title{Models model}
\usage{
measles_stan(n_cohorts, n_observations, ...)
}
\arguments{
\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}

\item{x}{Numeric vector of input values.}

\item{y}{Numeric vector of output values.}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Models model
}
