% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SEIR.R
\name{simulate_SEIR}
\alias{simulate_SEIR}
\title{Create SEIR model.}
\usage{
simulate_SEIR(
  pop = 1e+06,
  R0 = 15,
  vaccine_coverage = 0.9,
  SIA_day = 60,
  SIA_coverage = 0.7,
  gamma = 1/7,
  sigma = 1/10,
  days = 180
)
}
\arguments{
\item{pop}{Population.}
}
\description{
Create SEIR model.
}
