% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_stan.R
\name{measles_stan}
\alias{measles_stan}
\title{Models model}
\usage{
measles_stan(data, ...)
}
\arguments{
\item{data}{Data. TODO: add format.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Models model
}
